+++ vrp-core/src/construction/features/breaks.rs
@@ -186,7 +186,7 @@ struct OptionalBreakObjective {
 }
 
 impl FeatureObjective for OptionalBreakObjective {
-    fn fitness(&self, solution: &InsertionContext) -> f64 {
+    fn fitness(&self, solution: &InsertionContext) -> Cost {
         solution
             .solution
             .routes