+++ vrp-core/src/construction/features/total_value.rs
@@ -50,7 +50,7 @@ struct MaximizeTotalValueObjective {
 }
 
 impl FeatureObjective for MaximizeTotalValueObjective {
-    fn fitness(&self, solution: &InsertionContext) -> f64 {
+    fn fitness(&self, solution: &InsertionContext) -> Cost {
         solution.solution.routes.iter().fold(0., |acc, route_ctx| {
             route_ctx.route().tour.jobs().fold(acc, |acc, job| acc + (self.estimate_value_fn)(route_ctx, job))
         })